Final push. Broken but useful.
This was my first attempt to make the 'hello world' application have an
action bar. The attempt has broken the project, but that is fine: I
learned quite a bit about, and that was the point of the project. I will
quickly recreate a working base for the project and add an action bar to
it. LAST COMMIT.